requirements.txt
matplotlib
beautifulsoup4
requests
python-dotenv
